requirements.txt
nosetests
pycrypto
IPy
asn1
crytography
